requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
anyio==4.2.0
    # via httpx
beautifulsoup4==4.12.2
    # via facility-bot (setup.py)
blinker==1.7.0
    # via flask
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flask==3.0.0
    # via facility-bot (setup.py)
fonttools==4.47.0
    # via matplotlib
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via python-telegram-bot
icalendar==5.0.11
    # via facility-bot (setup.py)
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.8.2
    # via facility-bot (setup.py)
numpy==1.26.3
    # via
    #   contourpy
    #   matplotlib
packaging==23.2
    # via matplotlib
phonenumbers==8.13.27
    # via facility-bot (setup.py)
pillow==10.2.0
    # via matplotlib
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   facility-bot (setup.py)
    #   icalendar
    #   matplotlib
    #   vobject
python-telegram-bot==20.7
    # via facility-bot (setup.py)
pytz==2023.3.post1
    # via icalendar
pyzbar==0.1.9
    # via facility-bot (setup.py)
redis==5.0.1
    # via facility-bot (setup.py)
requests==2.31.0
    # via facility-bot (setup.py)
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
urllib3==2.1.0
    # via requests
vobject==0.9.6.1
    # via facility-bot (setup.py)
werkzeug==3.0.1
    # via flask
